requirements.txt
black @ file:///C:/b/abs_620t6ndje8/croot/black_1680737261963/work
brotlipy==0.7.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646056762388/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy==1.0.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
flit_core @ file:///C:/b/abs_eayrusf37u/croot/flit-core_1679397110577/work/source/flit_core
fonttools==4.39.3
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.27.0
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
mpmath==1.2.1
mypy-extensions==0.4.3
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
opencv-python==4.7.0.72
packaging==23.1
pathspec @ file:///C:/b/abs_9cu5_2yb3i/croot/pathspec_1674681579249/work
Pillow==9.4.0
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydicom==2.3.1
pylibjpeg==1.4.0
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
PyWavelets==1.4.1
requests @ file:///C:/b/abs_80_edxtvhu/croot/requests_1678709752318/work
scikit-image==0.20.0
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tifffile==2023.4.12
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
